+++ src/app/app.module.ts
@@ -4,9 +4,6 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatListModule } from '@angular/material/list';
-import { MatButtonModule } from '@angular/material/button';
 import { OrderModule } from '../app/order/order.module';
 
 @NgModule({
@@ -17,10 +14,7 @@ import { OrderModule } from '../app/order/order.module';
     OrderModule,
     BrowserModule,
     AppRoutingModule,
-    BrowserAnimationsModule,
-    MatCheckboxModule,
-    MatListModule,
-    MatButtonModule
+    BrowserAnimationsModule
   ],
   providers: [],
   bootstrap: [AppComponent]