+++ src/app/shared/shared.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HightlightDirective } from './directives/hightlight.directive';
+
+
+@NgModule({
+  declarations: [HightlightDirective],
+  imports: [
+    CommonModule
+  ],
+  exports: [HightlightDirective]
+})
+export class SharedModule { }