+++ changelog.md
@@ -1,5 +1,16 @@
+1/18/21
 Added:
 1. Display list of products
 2. Ability to buy product
 3. Display product in cart
 4. Remove product from cart
+
+1/27/21
+Added:
+1. Order module and component
+2. Communication service
+3. HighlightDirective 
+4. Cart item
+Changed:
+1. Cart list
+2. Cart service